+++ CHANGELOG.md
@@ -23,6 +23,8 @@ All notable changes to this project will be documented in this file.
 - **plugin-text**. Added headings plugin
 - **plugin-text**. Added list plugin
 - **plugin-text**. Added colors plugin
+- **plugin-text**. Added button to wrap in a blockquote plugin
+- **plugin-rows**. Pass `remove` and `replace` functions to childs
 
 ### Changed
 
@@ -32,6 +34,7 @@ All notable changes to this project will be documented in this file.
 - **ui** / **editor-ui** / **renderer-ui**. Revised theming workflow
 - **ui**. reworked the ui of several plugins (image, video, anchor, geogebra, h5p, highlight, hint, solution and spoiler)
 - **plugin-rows**. Improved styling of add menu and controls
+- **core**. Add new text plugin on Enter and delete empty plugins on backspace and delete
 
 ### Fixed